feat(components): Add CodeEditor component and basic functionality  Adds a CodeEditor component to the components folder. This might not be the optimal file structure yet and will be refactored later. Adds an onSelect handler to set the language based on the selected option from the constants.jsx file.
feat(components): Add CodeEditor component and basic functionality

Adds a CodeEditor component to the components folder. This might not be the optimal file structure yet and will be refactored later. Adds an onSelect handler to set the language based on the selected option from the constants.jsx file.
